% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bbox_to_polygon}
\alias{bbox_to_polygon}
\title{Convert a bounding box to a polygon in OGC WKT format.}
\usage{
bbox_to_polygon(bbox)
}
\arguments{
\item{bbox}{Numeric vector of length four containing xmin, ymin,
xmax, ymax.}
}
\value{
Character string for an OGC WKT polygon.
}
\description{
Returns a WKT polygon string for the given bounding box.
}
\examples{
elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
ds <- new(GDALRaster, elev_file, read_only=TRUE)
bbox_to_polygon(ds$bbox())
ds$close()
}
\seealso{
\code{\link[=bbox_from_wkt]{bbox_from_wkt()}}
}
