% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bbox_to_wkt}
\alias{bbox_to_wkt}
\title{Convert a bounding box to POLYGON in OGC WKT format.}
\usage{
bbox_to_wkt(bbox, x_ext = 0, y_ext = 0)
}
\arguments{
\item{bbox}{Numeric vector of length four containing xmin, ymin,
xmax, ymax.}

\item{x_ext}{Numeric scalar. Distance in \code{bbox} units to extend the
rectangle along the x-axis (results in \code{xmin = bbox[1] - x_ext},
\code{xmax = bbox[3] + x_ext}).}

\item{y_ext}{Numeric scalar. Distance in \code{bbox} units to extend the
rectangle along the y-axis (results in \code{ymin = bbox[2] - y_ext},
\code{ymax = bbox[4] + y_ext}).}
}
\value{
Character string for an OGC WKT polygon.
\code{NA} is returned if GDAL was built without the GEOS library.
}
\description{
\code{bbox_to_wkt()} returns a WKT POLYGON string for the given bounding box.
Requires GDAL built with the GEOS library.
}
\examples{
elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
ds <- new(GDALRaster, elev_file, read_only=TRUE)
bbox_to_wkt(ds$bbox())
ds$close()
}
\seealso{
\code{\link[=bbox_from_wkt]{bbox_from_wkt()}}, \code{\link[=g_buffer]{g_buffer()}}
}
