% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{buffer_wkt}
\alias{buffer_wkt}
\title{Compute buffer of a WKT geometry}
\usage{
buffer_wkt(geom, dist, quad_segs = 30)
}
\arguments{
\item{geom}{Character. OGC WKT string for a simple feature 2D geometry.}

\item{dist}{Numeric buffer distance in units of \code{geom}.}

\item{quad_segs}{Integer number of segments used to define a 90 degree
curve (quadrant of a circle). Large values result in large numbers of
vertices in the resulting buffer geometry while small numbers reduce the
accuracy of the result.}
}
\value{
Character string for an OGC WKT polygon.
\code{NA} is returned if GDAL was built without the GEOS library.
}
\description{
\code{buffer_wkt()} builds a new geometry containing the buffer region around
the geometry on which it is invoked. The buffer is a polygon containing
the region within the buffer distance of the original geometry.
}
\examples{
pt <- "POINT (0 0)"
bbox_from_wkt(buffer_wkt(geom = pt, dist = 10))
}
\seealso{
\code{\link[=bbox_from_wkt]{bbox_from_wkt()}}, \code{\link[=bbox_to_wkt]{bbox_to_wkt()}}
}
