% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalraster_proc.R
\docType{data}
\name{DEFAULT_DEM_PROC}
\alias{DEFAULT_DEM_PROC}
\title{List of default DEM processing options}
\format{
An object of class \code{list} of length 7.
}
\usage{
DEFAULT_DEM_PROC
}
\description{
These values are used in \code{dem_proc()} as the default processing options:
\preformatted{
    list(hillshade = c("-z", "1", "-s", "1", "-az", "315",
                       "-alt", "45", "-alg", "Horn",
                       "-combined", "-compute_edges"),
         slope = c("-s", "1", "-alg", "Horn", "-compute_edges"),
         aspect = c("-alg", "Horn", "-compute_edges"),
         color_relief = character(),
         TRI = c("-alg", "Riley", "-compute_edges"),
         TPI = c("-compute_edges"),
         roughness = c("-compute_edges"))
}
}
\seealso{
\code{\link[=dem_proc]{dem_proc()}}

\url{https://gdal.org/programs/gdaldem.html} for a description of all
available command-line options for each processing mode
}
\keyword{datasets}
