% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalraster_proc.R
\docType{data}
\name{DEFAULT_NODATA}
\alias{DEFAULT_NODATA}
\title{List of default nodata values by raster data type}
\format{
An object of class \code{list} of length 7.
}
\usage{
DEFAULT_NODATA
}
\description{
These values are currently used when a nodata value is needed but has not
been specified. Defined as:
\preformatted{
    list("Byte"= 255, "UInt16"= 65535, "Int16"= -32767,
         "UInt32"= 4294967293, "Int32"= -2147483647, 
         "Float32"= 3.402823466E+38, 
         "Float64"= .Machine$double.xmax)
}
}
\examples{
DEFAULT_NODATA[["UInt32"]]
}
\keyword{datasets}
