% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalraster_proc.R
\name{dem_proc}
\alias{dem_proc}
\title{GDAL DEM processing}
\usage{
dem_proc(
  mode,
  srcfile,
  dstfile,
  mode_options = DEFAULT_DEM_PROC[[mode]],
  color_text_file = NULL
)
}
\arguments{
\item{mode}{Character. Name of the DEM processing mode to use. One of}

\item{srcfile}{Filename of the the source elevation raster.}

\item{dstfile}{Filename of the output raster.}

\item{mode_options}{An optional character vector of command-line options
(see \link{DEFAULT_DEM_PROC}).}

\item{color_text_file}{Filename of a text file containing lines formatted
as: "elevation_value red green blue". Only used when \code{mode} is
\code{color-relief}.}
}
\value{
Logical indicating success (invisible \code{TRUE}).
An error is raised if the operation fails.
}
\description{
\code{dem_proc()} generates DEM derivatives from an input elevation raster. This
function is a wrapper for the \command{gdaldem} command-line utility.
See \url{https://gdal.org/programs/gdaldem.html} for details.
}
