% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{epsg_to_wkt}
\alias{epsg_to_wkt}
\title{Convert EPSG spatial reference to Well Known Text (WKT)}
\usage{
epsg_to_wkt(epsg, pretty = FALSE)
}
\arguments{
\item{epsg}{Integer EPSG code.}

\item{pretty}{Logical. \code{TRUE} to return a nicely formatted WKT 1 string
for display to a person. \code{FALSE} for a regular WKT string (the default).}
}
\value{
Character vector containing OGC WKT.
}
\description{
Exports the spatial reference for the specified EPSG code to WKT format.
As of GDAL 3.0, the default format for WKT export is OGC WKT 1.
}
\examples{
epsg_to_wkt(5070)
writeLines(epsg_to_wkt(5070, pretty=TRUE))
}
