% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_area}
\alias{g_area}
\title{Compute the area of a geometry}
\usage{
g_area(geom, quiet = FALSE)
}
\arguments{
\item{geom}{Either a raw vector of WKB or list of raw vectors, or a
character vector containing one or more WKT strings.}

\item{quiet}{Logical, \code{TRUE} to suppress warnings. Defaults to \code{FALSE}.}
}
\value{
Numeric vector of length equal to the number of input geometries.
Area or '0' if undefined.
}
\description{
\code{g_area()} computes the area for a \code{LinearRing}, \code{Polygon} or
\code{MultiPolygon}. Undefined for all other geometry types (returns zero).
}
\note{
\code{LinearRing} is a non-standard geometry type, used in GDAL just for geometry
creation.
}
