% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_centroid}
\alias{g_centroid}
\title{Compute the centroid of a geometry}
\usage{
g_centroid(geom, quiet = FALSE)
}
\arguments{
\item{geom}{Either a raw vector of WKB or list of raw vectors, or a
character vector containing one or more WKT strings.}

\item{quiet}{Logical, \code{TRUE} to suppress warnings. Defaults to \code{FALSE}.}
}
\value{
Either a numeric vector of length 2 containing the centroid (X, Y),
or a two-column numeric matrix (X, Y) with number of rows equal to the
number of input geometries.
}
\description{
\code{g_centroid()} returns a vector of point X, point Y, or a matrix of points
for multiple input geometries.
}
\details{
The GDAL documentation states "This method relates to the SFCOM
\code{ISurface::get_Centroid()} method however the current implementation based
on GEOS can operate on other geometry types such as multipoint, linestring,
geometrycollection such as multipolygons. OGC SF SQL 1.1 defines the
operation for surfaces (polygons). SQL/MM-Part 3 defines the operation for
surfaces and multisurfaces (multipolygons)."
}
