% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_contains}
\alias{g_contains}
\title{Determine whether a geometry contains the other geometry}
\usage{
g_contains(this_geom, other_geom)
}
\arguments{
\item{this_geom}{Character. OGC WKT string for a simple feature geometry.}

\item{other_geom}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
logical scalar. \code{TRUE} if \code{this_geom} contains \code{other_geom},
otherwise \code{FALSE}.
}
\description{
\code{g_contains()} tests if this geometry contains the other geometry.
}
\note{
Geometry validity is not checked. In case you are unsure of the validity
of the input geometries, call \code{g_is_valid()} before, otherwise the result
might be wrong.
}
