% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_difference}
\alias{g_difference}
\title{Generate a difference between two geometries}
\usage{
g_difference(this_geom, other_geom)
}
\arguments{
\item{this_geom}{Character. OGC WKT string for a simple feature geometry.}

\item{other_geom}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
Character string. Difference of the two geometries as OGC WKT.
}
\description{
\code{g_difference()} returns a new geometry which is the region of this geometry
with the region of the other geometry removed.
}
\note{
Geometry validity is not checked. In case you are unsure of the validity
of the input geometries, call \code{g_is_valid()} before, otherwise the result
might be wrong.
}
