% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_equals}
\alias{g_equals}
\title{Determine whether two geometries are equivalent}
\usage{
g_equals(this_geom, other_geom)
}
\arguments{
\item{this_geom}{Character. OGC WKT string for a simple feature geometry.}

\item{other_geom}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
logical scalar. \code{TRUE} if the two geometries are equivalent,
otherwise \code{FALSE}.
}
\description{
\code{g_equals()} returns \code{TRUE} if two geometries are equivalent.
}
\details{
The GDAL documentation says: "This operation implements the SQL/MM
\code{ST_OrderingEquals()} operation. The comparison is done in a structural way,
that is to say that the geometry types must be identical, as well as the
number and ordering of sub-geometries and vertices. Or equivalently, two
geometries are considered equal by this method if their WKT/WKB
representation is equal. Note: this must be distinguished from equality in
a spatial way."
}
