% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_export_to_json}
\alias{g_export_to_json}
\title{Convert WKB/WKT geometries into GeoJSON-style format}
\usage{
g_export_to_json(geom, srs = NULL, options = NULL)
}
\arguments{
\item{geom}{Either a raw vector of WKB or list of raw vectors, or a
character vector containing one or more WKT strings.}

\item{srs}{Optional character string specifying the spatial reference system
for the geometries given in \code{geom}. May be in WKT format or any of the
formats supported by \code{\link[=srs_to_wkt]{srs_to_wkt()}}.}

\item{options}{Optional character vector of \code{NAME=VALE} pairs. May include
any of the options supported by \code{OGR_G_ExportToJsonEx()} in the GDAL API
(see Details).}
}
\value{
A character vector of GeoJSON strings the same length as the number of input
geometries.
}
\description{
\code{g_export_to_json()} exports input geometries to GeoJSON strings. Interface
to \code{OGR_G_ExportToJsonEx()} in the GDAL API.
}
\details{
Available \code{options} are the ones supported by \code{OGR_G_ExportToJsonEx()}. See
\url{https://gdal.org/en/latest/api/vector_c_api.html}.

If there is a SRS attached to the geometry, and the geometry is aimed at
being stored in the "place" member of JSON-FG features, then
\code{COORDINATE_ORDER=AUTHORITY_COMPLIANT} option must be set (added in GDAL
3.12.1). When a SRS is attached to the geometry, and \code{AUTHORITY_COMPLIANT}
is used, the coordinates will be emitted in the order of the official SRS
definition. When using \code{TRADITIONAL_GIS_ORDER} (the default), coordinates
are emitted in longitude/easting first, then latitude/northing second. When
no SRS is attached, coordinates are emitted in the order they are set in the
geometry.
}
\examples{
g_export_to_json("POINT (-114.0 47.0)")
}
