% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_intersects}
\alias{g_intersects}
\title{Determine whether two geometries intersect}
\usage{
g_intersects(this_geom, other_geom)
}
\arguments{
\item{this_geom}{Character. OGC WKT string for a simple feature geometry.}

\item{other_geom}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
logical scalar. \code{TRUE} if the two geometries intersect, otherwise
\code{FALSE}.
}
\description{
\code{g_intersects()} tests whether two geometries intersect.
}
