% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_is_empty}
\alias{g_is_empty}
\title{Test if a geometry is empty}
\usage{
g_is_empty(geom, quiet = FALSE)
}
\arguments{
\item{geom}{Either a raw vector of WKB or list of raw vectors, or a
character vector containing one or more WKT strings.}

\item{quiet}{Logical, \code{TRUE} to suppress warnings. Defaults to \code{FALSE}.}
}
\value{
logical vector of the same length as the number of input
geometries in \code{geom}, containing \code{TRUE} for the corresponding geometries
that are empty or \code{FALSE} for non-empty geometries.
}
\description{
\code{g_is_empty()} tests whether a geometry has no points.
}
\examples{
g1 <- "POLYGON ((0 0, 10 10, 10 0, 0 0))"
g2 <- "POLYGON ((5 1, 9 5, 9 1, 5 1))"
g_difference(g2, g1) |> g_is_empty()
}
