% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_is_valid}
\alias{g_is_valid}
\title{Test if a geometry is valid}
\usage{
g_is_valid(geom, quiet = FALSE)
}
\arguments{
\item{geom}{Either a raw vector of WKB or list of raw vectors, or a
character vector containing one or more WKT strings.}

\item{quiet}{Logical, \code{TRUE} to suppress warnings. Defaults to \code{FALSE}.}
}
\value{
logical vector of the same length as the number of input
geometries in \code{geom}, containing \code{TRUE} for the corresponding geometries
that are valid or \code{FALSE} for invalid geometries.
}
\description{
\code{g_is_valid()} tests whether a geometry is valid.
}
\examples{
g1 <- "POLYGON ((0 0, 10 10, 10 0, 0 0))"
g2 <- "POLYGON ((0 0, 10 10, 10 0))"
g3 <- "POLYGON ((0 0, 10 10, 10 0, 0 1))"
g_is_valid(c(g1, g2, g3))
}
