% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_length}
\alias{g_length}
\title{Compute the length of a geometry}
\usage{
g_length(geom, quiet = FALSE)
}
\arguments{
\item{geom}{Either a raw vector of WKB or list of raw vectors, or a
character vector containing one or more WKT strings.}

\item{quiet}{Logical, \code{TRUE} to suppress warnings. Defaults to \code{FALSE}.}
}
\value{
Numeric vector of length equal to the number of input geometries.
Length or '0' if undefined.
}
\description{
\code{g_length()} computes the length for \code{LineString} or \code{MultiCurve} objects.
Undefined for all other geometry types (returns zero).
}
\examples{
g_length("LINESTRING (0 0, 3 4)")
}
