% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_name}
\alias{g_name}
\title{Extract geometry type names from WKB/WKT geometries}
\usage{
g_name(geom, quiet = FALSE)
}
\arguments{
\item{geom}{Either a raw vector of WKB or list of raw vectors, or a
character vector containing one or more WKT strings.}

\item{quiet}{Logical, \code{TRUE} to suppress warnings. Defaults to \code{FALSE}.}
}
\value{
character vector of the same length as the number of input
geometries in \code{geom}, containing the WKT names for the corresponding
geometries.
}
\description{
\code{g_name()} returns geometry type names in well known text format.
}
\examples{
elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
ds <- new(GDALRaster, elev_file)
bbox_to_wkt(ds$bbox()) |> g_name()
ds$close()
}
