% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_overlaps}
\alias{g_overlaps}
\title{Determine whether two geometries overlap}
\usage{
g_overlaps(this_geom, other_geom)
}
\arguments{
\item{this_geom}{Character. OGC WKT string for a simple feature geometry.}

\item{other_geom}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
logical scalar. \code{TRUE} if the two geometries overlap,
otherwise \code{FALSE}.
}
\description{
\code{g_overlaps()} tests if this geometry and the other geometry overlap, that
is, their intersection has a non-zero area (they have some but not all
points in common).
}
\note{
Geometry validity is not checked. In case you are unsure of the validity
of the input geometries, call \code{g_is_valid()} before, otherwise the result
might be wrong.
}
