% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_sym_difference}
\alias{g_sym_difference}
\title{Generate a symmetric difference between two geometries}
\usage{
g_sym_difference(this_geom, other_geom)
}
\arguments{
\item{this_geom}{Character. OGC WKT string for a simple feature geometry.}

\item{other_geom}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
Character string. Symmetric difference of the two geometries as
OGC WKT.
}
\description{
\code{g_sym_difference()} returns a new geometry which is the symmetric
difference of this geometry and the other geometry (union minus
intersection).
}
\note{
Geometry validity is not checked. In case you are unsure of the validity
of the input geometries, call \code{g_is_valid()} before, otherwise the result
might be wrong.
}
