% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_touches}
\alias{g_touches}
\title{Determine whether two geometries are touching}
\usage{
g_touches(this_geom, other_geom)
}
\arguments{
\item{this_geom}{Character. OGC WKT string for a simple feature geometry.}

\item{other_geom}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
logical scalar. \code{TRUE} if the two geometries are touching,
otherwise \code{FALSE}.
}
\description{
\code{g_touches()} tests if this geometry and the other geometry are touching.
}
\note{
Geometry validity is not checked. In case you are unsure of the validity
of the input geometries, call \code{g_is_valid()} before, otherwise the result
might be wrong.
}
