% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_union}
\alias{g_union}
\title{Generate a union of two geometries}
\usage{
g_union(this_geom, other_geom)
}
\arguments{
\item{this_geom}{Character. OGC WKT string for a simple feature geometry.}

\item{other_geom}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
Character string. Union of the two geometries as OGC WKT.
}
\description{
\code{g_union()} returns a new geometry which is the region of
union of the two geometries operated on.
}
\note{
Geometry validity is not checked. In case you are unsure of the validity
of the input geometries, call \code{g_is_valid()} before, otherwise the result
might be wrong.
}
