% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_within}
\alias{g_within}
\title{Determine whether a geometry is within the other geometry}
\usage{
g_within(this_geom, other_geom)
}
\arguments{
\item{this_geom}{Character. OGC WKT string for a simple feature geometry.}

\item{other_geom}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
logical scalar. \code{TRUE} if \code{this_geom} is within \code{other_geom},
otherwise \code{FALSE}.
}
\description{
\code{g_within()} tests if this geometry is within the other geometry.
}
\note{
Geometry validity is not checked. In case you are unsure of the validity
of the input geometries, call \code{g_is_valid()} before, otherwise the result
might be wrong.
}
