% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalmdimraster.R
\name{GDALMDIMRaster-class}
\alias{GDALMDIMRaster-class}
\alias{Rcpp_GDALMDIMRaster}
\alias{Rcpp_GDALMDIMRaster-class}
\alias{GDALMDIMRaster}
\title{Class wrapping the GDAL Multidim API}
\arguments{
\item{dsn}{Character string containing the filename to open. It may be
a file in a regular local filesystem, or a filename with a GDAL /vsiPREFIX/
(see \url{https://gdal.org/user/virtual_file_systems.html}).}
}
\description{
\code{GDALMDIMRaster} provides bindings to the GDAL Dataset API in MULTDIM mode.
}
