% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geos_version}
\alias{geos_version}
\title{Get GEOS version}
\usage{
geos_version()
}
\value{
A list of length four containing:
\itemize{
\item \code{name} - a string formatted as "major.minor.patch"
\item \code{major} - major version as integer
\item \code{minor} - minor version as integer
\item \code{patch} - patch version as integer
}

List elements will be \code{NA} if GDAL < 3.4.
}
\description{
\code{geos_version()} returns version information for the GEOS library in use by
GDAL. Requires GDAL >= 3.4.
}
\examples{
geos_version()
}
\seealso{
\code{\link[=gdal_version]{gdal_version()}}, \code{\link[=proj_version]{proj_version()}}
}
