% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_cache_used}
\alias{get_cache_used}
\title{Get the amount of memory currently in use by the GDAL block cache}
\usage{
get_cache_used()
}
\value{
Integer. Amount of memory in MB.
}
\description{
\code{get_cache_used()} returns the amount of memory in MB currently in use for
GDAL block caching.
}
\examples{
get_cache_used()
}
