% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_config_options}
\alias{get_config_options}
\title{Get all currently set GDAL configuration options}
\usage{
get_config_options()
}
\value{
A named character vector where names are configuration option
keys and values are the corresponding values. Returns an empty character
vector if no options are currently set.
}
\description{
\code{get_config_options()} returns all GDAL configuration options that have
been set via \code{set_config_option()} or other mechanism for setting GDAL
configuration options. This is useful for debugging to see
what options are currently active. Note that options set as environment
variables or via \code{CPLSetThreadLocalConfigOption()} are not included.
For a full description and listing of available options see
\url{https://gdal.org/en/stable/user/configoptions.html}.
}
\examples{
op <- get_config_option("GDAL_CACHEMAX")
set_config_option("GDAL_CACHEMAX", "12")
get_config_options()
set_config_option("GDAL_CACHEMAX", op)
}
\seealso{
\code{\link[=get_config_option]{get_config_option()}}, \code{\link[=set_config_option]{set_config_option()}}

\code{vignette("gdal-config-quick-ref")}
}
