% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_data_ptr}
\alias{get_data_ptr}
\title{Get pointer address of an R vector as a character string}
\usage{
get_data_ptr(x)
}
\arguments{
\item{x}{Vector of type \code{double}, \code{integer}, \code{raw} or \code{complex}.}
}
\value{
Character string pointer address with format suitable as
DATAPOINTER for a GDAL MEM dataset.
}
\description{
\code{get_data_ptr()} returns a character string representation of the address
of the first value in the C array underlying a given R vector of \code{raw},
\code{integer}, \code{double} or \code{complex}. The returned string is suitable for use
as a DATAPOINTER for a GDAL MEM dataset
(\url{https://gdal.org/en/stable/drivers/raster/mem.html}).
}
\examples{
v <- sample(0:255, 20, replace = TRUE)
get_data_ptr(v)
}
\seealso{
\code{\link[=rvector_to_MEM]{rvector_to_MEM()}}
}
