% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_cpus}
\alias{get_num_cpus}
\title{Get the number of processors detected by GDAL}
\usage{
get_num_cpus()
}
\value{
Integer scalar, number of CPUs.
}
\description{
\code{get_num_cpus()} returns the number of processors detected by GDAL.
Wrapper of \code{CPLGetNumCPUs()} in the GDAL Common Portability Library.
}
\examples{
get_num_cpus()
}
