% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_usable_physical_ram}
\alias{get_usable_physical_ram}
\title{Get usable physical RAM}
\usage{
get_usable_physical_ram()
}
\value{
Numeric scalar, bytes as \code{bit64::integer64} type.
}
\description{
\code{get_usable_physical_ram()} returns the total physical RAM, usable by a
process, in bytes.
}
\examples{
get_usable_physical_ram()
}
