% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_helpers.R
\name{getCreationOptions}
\alias{getCreationOptions}
\title{Return the list of creation options of a GDAL driver}
\usage{
getCreationOptions(format, filter = NULL)
}
\arguments{
\item{format}{Raster format short name (e.g., "GTiff").}

\item{filter}{Optional character vector of creation option names.}
}
\value{
List ... TODO
}
\description{
\code{getCreationOptions()} returns the list of creation options supported by a
GDAL format driver.
Wrapper for \code{GDALGetDriverCreationOptionList()} in the GDAL API.
}
\examples{
getCreationOptions("GTiff", filter="COMPRESS")
}
\seealso{
\code{\link[=GDALRaster]{GDALRaster-class}}, \code{\link[=create]{create()}}, \code{\link[=createCopy]{createCopy()}}
}
