% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{http_enabled}
\alias{http_enabled}
\title{Check if GDAL CPLHTTP services can be useful (libcurl)}
\usage{
http_enabled()
}
\value{
Logical scalar, \code{TRUE} if GDAL was built with \code{libcurl} support.
}
\description{
\code{http_enabled()} returns \code{TRUE} if \code{libcurl} support is enabled.
Wrapper of \code{CPLHTTPEnabled()} in the GDAL Common Portability Library.
}
\examples{
http_enabled()
}
