% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_helpers.R
\name{inspectDataset}
\alias{inspectDataset}
\title{Obtain information about a GDAL raster or vector dataset}
\usage{
inspectDataset(filename, ...)
}
\description{
Obtain information about a GDAL raster or vector dataset
}
