% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_helpers.R
\name{lib_versions}
\alias{lib_versions}
\title{Return library version information for GDAL and its dependencies}
\usage{
lib_versions()
}
\value{
A named list with elements \code{"gdal"}, \code{"proj"} and \code{"geos"}, each
containing a named list with the following elements:
\itemize{
\item \code{"name"}: character string version as \code{"major.minor.patch"}
\item \code{"major"}: integer major version number
\item \code{"minor"}: integer minor version number
\item \code{"patch"}: integer patch version number
}
}
\description{
\code{lib_versions()} returns a named list of library version information for
GDAL and its major dependencies, currently PROJ and GEOS. It provides library
versions in a consistent format, as an alternative to the separate
\code{gdal_version()}, \code{proj_version()} and \code{geos_version()}.
}
\examples{
lib_versions()
}
\seealso{
\code{\link[=gdal_version]{gdal_version()}}, \code{\link[=proj_version]{proj_version()}}, \code{\link[=geos_version]{geos_version()}}
}
