% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_mdim-2.R
\name{mdim_read_array_attrs}
\alias{mdim_read_array_attrs}
\title{Read attributes from a multidimensional array}
\usage{
mdim_read_array_attrs(dsn, array_name, open_options = NULL)
}
\arguments{
\item{dsn}{Character string containing the data source name.}

\item{array_name}{Character string giving the name of the array.}

\item{open_options}{Optional character vector of \code{NAME=VALUE} pairs
specifying dataset open options.}
}
\value{
A named list of attribute values.
}
\description{
\code{mdim_read_array_attrs()} reads all attributes from a named array,
returning a named list. Each attribute is auto-typed to the appropriate
R type (character, numeric, integer, or raw for compound types).
}
\examples{
\dontshow{if (gdal_version_num() >= gdal_compute_version(3, 1, 0) && isTRUE(gdal_formats("netCDF")$multidim_raster)) withAutoprint(\{ # examplesIf}
f <- system.file("extdata/byte.nc", package="gdalraster")

mdim_read_array_attrs(f, "Band1")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=mdim_read_group_attrs]{mdim_read_group_attrs()}}, \code{\link[=mdim_read_info]{mdim_read_info()}}
}
