% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_mdim-2.R
\name{mdim_read_coord_info}
\alias{mdim_read_coord_info}
\title{Get coordinate metadata for a dimension}
\usage{
mdim_read_coord_info(dsn, array_name, dim_index, open_options = NULL)
}
\arguments{
\item{dsn}{Character string containing the data source name.}

\item{array_name}{Character string giving the name of the array.}

\item{dim_index}{Integer. The 0-based index of the dimension.}

\item{open_options}{Optional character vector of \code{NAME=VALUE} pairs
specifying dataset open options.}
}
\value{
A list with components: \code{name}, \code{size}, \code{type}, \code{direction},
\code{has_coord_var} (logical), \code{units}, \code{calendar}. The \code{units} and
\code{calendar} components are \code{NULL} if not available.
}
\description{
\code{mdim_read_coord_info()} returns metadata about the coordinate
(indexing) variable associated with a dimension, including its name,
size, type, direction, units, and CF calendar attribute.
}
\examples{
\dontshow{if (gdal_version_num() >= gdal_compute_version(3, 1, 0) && isTRUE(gdal_formats("netCDF")$multidim_raster)) withAutoprint(\{ # examplesIf}
f <- system.file("extdata/byte.nc", package="gdalraster")

mdim_read_coord_info(f, "Band1", 0)
mdim_read_coord_info(f, "Band1", 1)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=mdim_read_dim_values]{mdim_read_dim_values()}}, \code{\link[=mdim_read_info]{mdim_read_info()}}
}
