% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_mdim-2.R
\name{mdim_read_dim_values}
\alias{mdim_read_dim_values}
\title{Read coordinate values for a dimension}
\usage{
mdim_read_dim_values(dsn, array_name, dim_index, open_options = NULL)
}
\arguments{
\item{dsn}{Character string containing the data source name.}

\item{array_name}{Character string giving the name of the array.}

\item{dim_index}{Integer. The 0-based index of the dimension.}

\item{open_options}{Optional character vector of \code{NAME=VALUE} pairs
specifying dataset open options.}
}
\value{
A numeric vector of coordinate values.
}
\description{
\code{mdim_read_dim_values()} reads the coordinate (indexing) variable values
for a specified dimension of a multidimensional array. If the dimension
has no associated coordinate variable, a 0-based integer sequence is
returned.
}
\examples{
\dontshow{if (gdal_version_num() >= gdal_compute_version(3, 1, 0) && isTRUE(gdal_formats("netCDF")$multidim_raster)) withAutoprint(\{ # examplesIf}
f <- system.file("extdata/byte.nc", package="gdalraster")

## read x coordinates (dimension 1)
mdim_read_dim_values(f, "Band1", 1)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=mdim_read]{mdim_read()}}, \code{\link[=mdim_read_coord_info]{mdim_read_coord_info()}}
}
