% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_mdim-2.R
\name{mdim_read_group_attrs}
\alias{mdim_read_group_attrs}
\title{Read attributes from the root group}
\usage{
mdim_read_group_attrs(dsn, group_name = NULL, open_options = NULL)
}
\arguments{
\item{dsn}{Character string containing the data source name.}

\item{group_name}{Optional character string giving the name of a
subgroup. Default (\code{NULL}) reads attributes from the root group.}

\item{open_options}{Optional character vector of \code{NAME=VALUE} pairs
specifying dataset open options.}
}
\value{
A named list of attribute values.
}
\description{
\code{mdim_read_group_attrs()} reads all attributes from the root group
(or a named subgroup) of a multidimensional raster dataset.
}
\examples{
\dontshow{if (gdal_version_num() >= gdal_compute_version(3, 1, 0) && isTRUE(gdal_formats("netCDF")$multidim_raster)) withAutoprint(\{ # examplesIf}
f <- system.file("extdata/byte.nc", package="gdalraster")

mdim_read_group_attrs(f)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=mdim_read_array_attrs]{mdim_read_array_attrs()}}, \code{\link[=mdim_info]{mdim_info()}}
}
