% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_mdim-2.R
\name{mdim_read_info}
\alias{mdim_read_info}
\title{Get information about a multidimensional array}
\usage{
mdim_read_info(dsn, array_name, open_options = NULL)
}
\arguments{
\item{dsn}{Character string containing the data source name.}

\item{array_name}{Character string giving the name of the array.}

\item{open_options}{Optional character vector of \code{NAME=VALUE} pairs
specifying dataset open options.}
}
\value{
A list with components: \code{name}, \code{fullname}, \code{dims} (list of
dimension info), \code{dim_names}, \code{shape}, \code{dtype}, \code{unit}, \code{nodata},
\code{scale}, \code{offset}, \code{spatial_ref}.
}
\description{
\code{mdim_read_info()} returns metadata about a named array in a
multidimensional raster dataset, including dimension names, shape,
data type, CF convention parameters, and spatial reference.
}
\examples{
\dontshow{if (gdal_version_num() >= gdal_compute_version(3, 1, 0) && isTRUE(gdal_formats("netCDF")$multidim_raster)) withAutoprint(\{ # examplesIf}
f <- system.file("extdata/byte.nc", package="gdalraster")

mdim_read_info(f, "Band1")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=mdim_read]{mdim_read()}}, \code{\link[=mdim_info]{mdim_info()}}
}
