% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_define.R
\name{ogr_define}
\alias{ogr_define}
\alias{ogr_def_field}
\alias{ogr_def_geom_field}
\title{OGR feature class definition for vector data}
\usage{
ogr_def_field(
  fld_type,
  subtype = NULL,
  width = NULL,
  precision = NULL,
  is_nullable = NULL,
  is_unique = NULL,
  is_ignored = NULL,
  default = NULL
)

ogr_def_geom_field(
  geom_type,
  srs = NULL,
  is_nullable = NULL,
  is_ignored = NULL
)
}
\description{
This topic contains documentation and helper functions for defining an
OGR feature class.
\code{ogr_def_field()} creates an attribute field definition, a list
containing the field data type and potentially other optional field
properties.
\code{ogr_def_geom_field()} similarly creates a geometry field definition.
A list containing zero or more attribute field definitions, along with one
or more geomtery field definitions, comprise an OGR feature class definition
(a.k.a. layer definition).
}
\details{
All features in an OGR Layer share a common schema (feature class), modeled
in GDAL as OGR Feature Definition. The feature class definition includes the
set of attribute fields and their data types and the geometry field(s).
In R, a feature class definition is represented as a list, having as names
the attribute/geometry field names, with each list element holding a field
definition.

An attribute field definition is a list with named elements:

\if{html}{\out{<div class="sourceCode">}}\preformatted{$type       : OGR Field Type ("OFTReal", "OFTString" etc.)
$subtype    : optional ("OFSTBoolean", ...)
$width      : optional max number of characters
$precision  : optional number of digits after the decimal point
$is_nullable: optional NOT NULL constraint (logical scalar)
$is_unique  : optional UNIQUE constraint (logical scalar)
$default    : optional default value as character string
$is_ignored : optionally ignored when retrieving features (logical scalar)
$is_geom    : FALSE (the default) for attribute fields
}\if{html}{\out{</div>}}

An OGR field type is specified as a character string with possible values:
\code{OFTInteger}, \code{OFTIntegerList}, \code{OFTReal}, \code{OFTRealList}, \code{OFTString},
\code{OFTStringList}, \code{OFTBinary},  \code{OFTDate}, \code{OFTTime}, \code{OFTDateTime},
\code{OFTInteger64}, \code{OFTInteger64List}.

An optional field subtype is specified as a character string with possible
values:
\code{OFSTNone}, \code{OFSTBoolean}, \code{OFSTInt16}, \code{OFSTFloat32}, \code{OFSTJSON},
\code{OFSTUUID}.

By default, fields are nullable, have no unique constraint, and are not
ignored (i.e., not omitted when fetching features). Not-null and unique
constraints are not supported by all format drivers.

A default field value is taken into account by format drivers (generally
those with a SQL interface) that support it at field creation time.
If given in the field definition, \verb{$default} must be a character string.
The accepted values are \code{"NULL"}, a numeric value (e.g., \code{"0"}), a literal
value enclosed between single quote characters (e.g., \code{"'a default value'"},
with any inner single quote characters escaped by repetition of the single
quote character), \code{"CURRENT_TIMESTAMP"}, \code{"CURRENT_TIME"}, \code{"CURRENT_DATE"}
or a driver-specific expression (that might be ignored by other drivers).
For a datetime literal value, format should be
\code{"'YYYY/MM/DD HH:MM:SS[.sss]'"} (considered as UTC time).

A geometry field definition is a list with named elements:

\if{html}{\out{<div class="sourceCode">}}\preformatted{$type       : geom type ("Point", "Polygon", etc.)
$srs        : optional spatial reference as WKT string
$is_nullable: optional NOT NULL constraint (logical scalar)
$is_ignored : optionally ignored when retrieving features (logical scalar)
$is_geom    : TRUE (required) for geometry fields
}\if{html}{\out{</div>}}

Typically, there is one geometry field on a layer, but some formats support
more than one geometry column per table (e.g., PostGIS).

Geometry types are specified as a character string containing OGC WKT.
Common types include: \code{Point}, \code{LineString}, \code{Polygon}, \code{MultiPoint},
\code{MultiLineString}, \code{MultiPolygon}. See the GDAL documentation for a list
of all supported geometry types:\cr
\url{https://gdal.org/api/vector_c_api.html#_CPPv418OGRwkbGeometryType}

Format drivers may or may not support not-null constraints on attribute and
geometry fields. If they support creating fields with not-null constraints,
this is generally before creating any features to the layer. In some cases,
a not-null constraint may be available as a layer creation option. For
example, GeoPackage format has a layer creation option
\verb{GEOMETRY_NULLABLE=[YES/NO]}.
}
\seealso{
the \link{ogr_manage} functions, \code{\link[=ogrinfo]{ogrinfo()}}

WKT representation of geometry:\cr
\url{https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry}
}
