% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{plot.OGRFeatureSet}
\alias{plot.OGRFeatureSet}
\title{Plot the geometry column of an \code{OGRFeatureSet}}
\usage{
\method{plot}{OGRFeatureSet}(x, ...)
}
\arguments{
\item{x}{An \code{OGRFeatureSet}.}

\item{...}{Optional arguments passed to \code{wk::wk_plot()}.}
}
\value{
The input, invisibly.
}
\description{
Plot the geometry column of an \code{OGRFeatureSet}
}
