% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{plot_raster_data}
\alias{plot_raster_data}
\title{Display raster data that has been read into a vector}
\usage{
plot_raster_data(
  data,
  xsize,
  ysize,
  nbands = 1,
  normalize = TRUE,
  col_scale_fn = NULL,
  xlim = c(0, xsize),
  ylim = c(ysize, 0),
  xlab = "x",
  ylab = "y",
  interpolate = TRUE,
  asp = 1,
  axes = TRUE,
  main = "",
  xaxs = "i",
  yaxs = "i",
  col_na = rgb(0, 0, 0, 0),
  ...
)
}
\arguments{
\item{data}{A numeric vector of pixel data to display, arranged in left to
right, top to bottom pixel order.}

\item{xsize}{The number of pixels along the x dimension in \code{data}.}

\item{ysize}{The number of pixels along the y dimension in \code{data}.}

\item{nbands}{The number of bands in \code{data}. Must be either 1 (grayscale) or
3 (RGB). For RGB, \code{data} are interleaved by band.}

\item{normalize}{Logical. \code{TRUE} to rescale pixel values so that their
range is \verb{[0,1}].}

\item{col_scale_fn}{An optional color scale function (default is \code{gray} for
single-band data or \code{rgb} for 3-band).}

\item{xlim}{Numeric vector of length two giving the x coordinate range in
pixel/line space. The default is \code{c(0, xsize)}.}

\item{ylim}{Numeric vector of length two giving the y coordinate range in
pixel/line space. The default is \code{c(ysize, 0)}.}

\item{xlab}{Title for the x axis (see \code{?title}).}

\item{ylab}{Title for the y axis (see \code{?title}).}

\item{interpolate}{Logical indicating whether to apply linear interpolation
to the image when drawing (default \code{TRUE}).}

\item{asp}{Numeric. The aspect ratio y/x (see \code{?plot.window}).}

\item{axes}{Logical. \code{TRUE} to draw axes (the default).}

\item{main}{The main title (on top).}

\item{xaxs}{The style of axis interval calculation to be used for the x axis
(see \code{?par}).}

\item{yaxs}{The style of axis interval calculation to be used for the y axis
(see \code{?par}).}

\item{col_na}{Color to use for \code{NA} as a 7 or 9-character hexadecimal code.
The default is transparent (\code{"#00000000"}, the return value of
\code{rgb(0,0,0,0)}).}

\item{...}{Other parameters to be passed to \code{plot.default}.}
}
\description{
\code{plot_raster_data()} displays raster data using base graphics.
}
