% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pop_error_handler}
\alias{pop_error_handler}
\title{Pop error handler off stack}
\usage{
pop_error_handler()
}
\value{
No return value, called for side effects.
}
\description{
\code{pop_error_handler()} is a wrapper for \code{CPLPopErrorHandler()} in the GDAL
Common Portability Library.
Discards the current error handler on the error handler stack, and restores
the one in use before the last \code{push_error_handler()} call. This method has
no effect if there are no error handlers on the current thread's error
handler stack.
}
\examples{
push_error_handler("quiet")
# ...
pop_error_handler()
}
\seealso{
\code{\link[=push_error_handler]{push_error_handler()}}
}
