% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.OGRFeatureSet}
\alias{print.OGRFeatureSet}
\title{Print an \code{OGRFeatureSet}}
\usage{
\method{print}{OGRFeatureSet}(x, ...)
}
\arguments{
\item{x}{An \code{OGRFeatureSet}.}

\item{...}{Optional arguments passed to \code{base::print.data.frame()}.}
}
\value{
The input, invisibly.
}
\description{
Print an \code{OGRFeatureSet}
}
