% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_helpers.R
\name{progress_bar_clear}
\alias{progress_bar_clear}
\title{Clear progress bar}
\usage{
progress_bar_clear()
}
\value{
No return value, call for side effects.
}
\description{
\code{progress_bar_clear()} terminates any active \pkg{cli} progress bars and
resets the global progress bar in C++. Generally not needed unless a process
using a progress bar terminates abnormally, or with ctrl-c interrupt, and
a progress bar display anomaly results.
}
