% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_version}
\alias{proj_version}
\title{Get PROJ version}
\usage{
proj_version()
}
\value{
A list of length four containing:
\itemize{
\item \code{name} - a string formatted as "major.minor.patch"
\item \code{major} - major version as integer
\item \code{minor} - minor version as integer
\item \code{patch} - patch version as integer
}

List elements will be \code{NA} if GDAL < 3.0.1.
}
\description{
\code{proj_version()} returns version information for the PROJ library in use by
GDAL. Requires GDAL >= 3.0.1.
}
\examples{
proj_version()
}
\seealso{
\code{\link[=gdal_version]{gdal_version()}}, \code{\link[=geos_version]{geos_version()}}, \code{\link[=proj_search_paths]{proj_search_paths()}},
\code{\link[=proj_networking]{proj_networking()}}
}
