% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{push_error_handler}
\alias{push_error_handler}
\title{Push a CPLQuietErrorHandler}
\usage{
push_error_handler(handler = "quiet")
}
\arguments{
\item{handler}{Character name of the error handler to push. \code{quiet} is
currently the only supported handler.}
}
\value{
No return value, called for side effects.
}
\description{
\code{push_error_handler()} is a wrapper for
\code{CPLPushErrorHandler(CPLQuietErrorHandler)} in the GDAL Common Portability
Library.
This pushes a new error handler on the thread-local error handler stack.
This handler will be used until removed with \code{pop_error_handler()}.
\code{CPLQuietErrorHandler()} doesn't make any attempt to report passed error or
warning messages but will process debug messages via
\code{CPLDefaultErrorHandler}.
}
\examples{
result <- deleteDataset("/vsimem/nonexistent.tif")
push_error_handler()
result <- deleteDataset("/vsimem/nonexistent.tif")
pop_error_handler()
}
\seealso{
\code{\link[=pop_error_handler]{pop_error_handler()}}
}
