% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vsi_clear_path_options}
\alias{vsi_clear_path_options}
\title{Clear path specific configuration options}
\usage{
vsi_clear_path_options(path_prefix)
}
\arguments{
\item{path_prefix}{Character string. If set to \code{""} (empty string), all
path specific options are cleared. If set to a path prefix, only those
options set with \code{vsi_set_path_option(path_prefix, ...)} will be cleared.}
}
\value{
No return value, called for side effect.
}
\description{
\code{vsi_clear_path_options()} clears path specific options previously set
with \code{vsi_set_path_option()}.
Wrapper for \code{VSIClearPathSpecificOptions()} in the GDAL Common Portability
Library. Requires GDAL >= 3.6.
}
\note{
No particular care is taken to remove options from RAM in a secure way.
}
\seealso{
\code{\link[=vsi_set_path_option]{vsi_set_path_option()}}
}
