% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsifile.R
\name{vsi_constants}
\alias{vsi_constants}
\alias{SEEK_SET}
\alias{SEEK_CUR}
\alias{SEEK_END}
\title{Constants for VSIFile$seek()}
\usage{
SEEK_SET

SEEK_CUR

SEEK_END
}
\description{
These are package global constants for convenience in calling
\code{VSIFile$seek()}.
}
