% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vsi_get_fs_prefixes}
\alias{vsi_get_fs_prefixes}
\title{Return the list of virtual file system handlers currently registered}
\usage{
vsi_get_fs_prefixes()
}
\value{
Character vector containing prefixes of the virtual file system
handlers.
}
\description{
\code{vsi_get_fs_prefixes()} returns the list of prefixes for virtual file
system handlers currently registered (e.g., \code{"/vsimem/"}, \code{"/vsicurl/"},
etc). Wrapper for \code{VSIGetFileSystemsPrefixes()} in the GDAL API.
}
\examples{
vsi_get_fs_prefixes()
}
\seealso{
\code{\link[=vsi_get_fs_options]{vsi_get_fs_options()}}

\url{https://gdal.org/en/stable/user/virtual_file_systems.html}
}
