% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vsi_uri_to_vsi_path}
\alias{vsi_uri_to_vsi_path}
\title{Return VSI compatible paths from URIs / URLs}
\usage{
vsi_uri_to_vsi_path(uris)
}
\arguments{
\item{uris}{Character vector of input URI strings.}
}
\value{
Character vector of VSI paths.
}
\description{
\code{vsi_uri_to_vsi_path()} substitutes URIs / URLs starting with \verb{s3://},
\verb{gs://}, etc. by their VSI prefix equivalents. If no known substitution is
found, the input string is returned unmodified.
Wrapper of \code{VSIURIToVSIPath()} in the GDAL API. Requires GDAL >= 3.12.
}
\examples{
\dontshow{if (gdal_version_num() >= gdal_compute_version(3, 12, 0)) withAutoprint(\{ # examplesIf}
# Requires GDAL >= 3.12
vsi_uri_to_vsi_path("gs://cmip6/")
\dontshow{\}) # examplesIf}
}
