% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalraster_proc.R
\docType{data}
\name{DEFAULT_NODATA}
\alias{DEFAULT_NODATA}
\title{List of default nodata values by raster data type}
\format{
An object of class \code{list} of length 8.
}
\usage{
DEFAULT_NODATA
}
\description{
These values are currently used in \code{gdalraster} when a nodata value is
needed but has not been specified:
\preformatted{
    list("Byte" = 255, "Int8" = -128,
         "UInt16" = 65535, "Int16" = -32767,
         "UInt32" = 4294967293, "Int32" = -2147483647, 
         "Float32" = -99999.0, "Float64" = -99999.0)
}
}
\keyword{datasets}
