% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_area}
\alias{g_area}
\title{Compute the area of a geometry}
\usage{
g_area(wkt)
}
\arguments{
\item{wkt}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
Numeric scalar. Area of the geometry or \code{0}.
}
\description{
\code{g_area()} computes the area for a \code{LinearRing}, \code{Polygon} or
\code{MultiPolygon}. Undefined for all other geometry types (returns zero).
}
\note{
\code{LinearRing} is a non-standard geometry type, used in GDAL just for geometry
creation.
}
