% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_centroid}
\alias{g_centroid}
\title{Compute the centroid of a geometry}
\usage{
g_centroid(wkt)
}
\arguments{
\item{wkt}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
Numeric vector of length 2 containing the centroid (X, Y).
}
\description{
\code{g_centroid()} returns a vector of point X, point Y.
}
\details{
The GDAL documentation states "This method relates to the SFCOM
\code{ISurface::get_Centroid()} method however the current implementation based
on GEOS can operate on other geometry types such as multipoint, linestring,
geometrycollection such as multipolygons. OGC SF SQL 1.1 defines the
operation for surfaces (polygons). SQL/MM-Part 3 defines the operation for
surfaces and multisurfaces (multipolygons)."
}
