% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_is_empty}
\alias{g_is_empty}
\title{Test if a geometry is empty}
\usage{
g_is_empty(wkt)
}
\arguments{
\item{wkt}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
logical scalar. \code{TRUE} if the geometry has no points, otherwise
\code{FALSE}.
}
\description{
\code{g_is_empty()} tests whether a geometry has no points.
}
\examples{
g1 <- "POLYGON ((0 0, 10 10, 10 0, 0 0))"
g2 <- "POLYGON ((5 1, 9 5, 9 1, 5 1))"
g_difference(g2, g1) |> g_is_empty()
}
