% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_is_valid}
\alias{g_is_valid}
\title{Test if a geometry is valid}
\usage{
g_is_valid(wkt)
}
\arguments{
\item{wkt}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
logical scalar. \code{TRUE} if the geometry is valid, otherwise
\code{FALSE}.
}
\description{
\code{g_is_valid()} tests whether a geometry is valid.
}
\examples{
g1 <- "POLYGON ((0 0, 10 10, 10 0, 0 0))"
g_is_valid(g1)

g2 <- "POLYGON ((0 0, 10 10, 10 0, 0 1))"
g_is_valid(g2)
}
