% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g_length}
\alias{g_length}
\title{Compute the length of a geometry}
\usage{
g_length(wkt)
}
\arguments{
\item{wkt}{Character. OGC WKT string for a simple feature geometry.}
}
\value{
Numeric scalar. Length of the geometry or \code{0}.
}
\description{
\code{g_length()} computes the length for \code{LineString} or \code{MultiCurve} objects.
Undefined for all other geometry types (returns zero).
}
\examples{
g_length("LINESTRING (0 0, 3 4)")
}
