% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdal_formats}
\alias{gdal_formats}
\title{Report all configured GDAL drivers for raster formats}
\usage{
gdal_formats()
}
\value{
No return value, called for reporting only.
}
\description{
\code{gdal_formats()} prints to the console a list of the supported raster
formats.
}
\examples{
gdal_formats()
}
