% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdal_version}
\alias{gdal_version}
\title{Get GDAL version}
\usage{
gdal_version()
}
\value{
Character vector of length four containing:
\itemize{
\item "–version" - one line version message, e.g., “GDAL 3.6.3, released
2023/03/12”
\item "GDAL_VERSION_NUM" - formatted as a string, e.g., “3060300” for
GDAL 3.6.3.0
\item "GDAL_RELEASE_DATE" - formatted as a string, e.g., “20230312”
\item "GDAL_RELEASE_NAME" - e.g., “3.6.3”
}
}
\description{
\code{gdal_version()} returns runtime version information.
}
\examples{
gdal_version()
}
