% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalalg.R
\name{GDALAlg-class}
\alias{GDALAlg-class}
\alias{Rcpp_GDALAlg}
\alias{Rcpp_GDALAlg-class}
\alias{GDALAlg}
\title{Class encapsulating a GDAL CLI Algorithm}
\description{
\code{GDALAlg} provides an interface for accessing...
}
