% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_data_ptr}
\alias{get_data_ptr}
\title{Get pointer address of R data as character string}
\usage{
get_data_ptr(x)
}
\arguments{
\item{x}{Object of type numeric, integer or raw.}
}
\value{
Character string pointer address  with format suitable for use
as DATAPOINTER for GDAL MEM.
}
\description{
Get pointer address of R data as character string
}
