% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{has_geos}
\alias{has_geos}
\title{Is GEOS available?}
\usage{
has_geos()
}
\value{
Logical. \code{TRUE} if GEOS is available, otherwise \code{FALSE}.
}
\description{
\code{has_geos()} returns a logical value indicating whether GDAL was built
against the GEOS library.
}
\examples{
has_geos()
}
