% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdim_translate}
\alias{mdim_translate}
\title{Convert multidimensional data between different formats}
\usage{
mdim_translate(
  src_dsn,
  dst_dsn,
  output_format = "",
  creation_options = NULL,
  array_specs = NULL,
  group_specs = NULL,
  subset_specs = NULL,
  scaleaxes_specs = NULL,
  allowed_drivers = NULL,
  open_options = NULL,
  strict = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{src_dsn}{Character string giving the name of the source
multidimensional raster dataset (e.g., file, VSI path).}

\item{dst_dsn}{Character string giving the name of the destination
multidimensional raster dataset (e.g., file, VSI path).}

\item{output_format}{Character string giving the output format (driver short
name). This can be a format that supports multidimensional output (such as
NetCDF: Network Common Data Form, Multidimensional VRT), or a "classic" 2D
format, if only one single 2D array results from the other specified
conversion operations. When this option is not specified (i.e., empty string
\code{""}), the format is guessed when possible from the extension of \code{dst_dsn}.}

\item{creation_options}{Optional character vector of format-specific
creation options as \code{"NAME=VALUE"} pairs. A list of options supported for a
format can be obtained with \code{getCreationOptions()}, but the documentation
for the format is the definitive source of information on driver creation
options (see \url{https://gdal.org/en/stable/drivers/raster/index.html}).
Array-level creation options may be passed by prefixing them with \verb{ARRAY:}.}

\item{array_specs}{Optional character vector of one or more array
specifications, instead of coverting the whole dataset (see Details).}

\item{group_specs}{Optional character vector of one or more array
specifications, instead of coverting the whole dataset (see Details).}

\item{subset_specs}{Optional character vector of one or more subset
specifications, that perform trimming or slicing along a dimension, provided
that it is indexed by a 1D variable of numeric or string data type, and
whose values are monotonically sorted (see Details).}

\item{scaleaxes_specs}{Optional character vector of one or more scale axes
specifications, that apply an integral scale factor to one or several
dimensions, that is extract 1 value every N values (without resampling) (see
Details).}

\item{open_options}{Optional character vector of format-specific dataset open
options as \code{"NAME=VALUE"} pairs.}

\item{strict}{Logical value, \code{FALSE} (the default) some failures during the
translation are tolerated, such as not being able to write group attributes.
If set to \code{TRUE}, such failures will cause the process to fail.}

\item{quiet}{Logical value, set to \code{TRUE} to disable progress reporting.
Defaults to \code{FALSE}.}
}
\value{
Logical value indicating success (invisible \code{TRUE}, output written to
\code{dst_dsn}). An error is raised if the operation fails.
}
\description{
\code{mdim_translate()} is a wrapper of the \command{gdalmdimtranslate}
command-line utility (see
\url{https://gdal.org/en/stable/programs/gdalmdimtranslate.html}).
This function converts multidimensional data between different formats and
performs subsetting. Requires GDAL >= 3.2.
}
\details{
\subsection{\code{array_specs}}{
Instead of converting the whole dataset, select one or more arrays, and
possibly perform operations on them. One or more array specifications can
be given as elements of a character vector.

An array specification may be just an array name, potentially using a fully
qualified syntax (\code{"/group/subgroup/array_name"}). Or it can be a
combination of options with the syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{name=\{src_array_name\}[,dstname=\{dst_array_name\}][,resample=yes][,transpose=[\{axis1\},\{axis2\},...][,view=\{view_expr\}]
}\if{html}{\out{</div>}}

The following options are processed in that order:
\itemize{
\item \code{resample=yes} asks for the array to run through
\code{GDALMDArray::GetResampled()}.
\item \verb{[\{axis1\},\{axis2\},...]} is the argument of \code{GDALMDArray::Transpose()}. For
example, \verb{transpose=[1,0]} switches the axis order of a 2D array.
\item \code{{view_expr}} is the value of the \code{viewExpr} argument of
\code{GDALMDArray::GetView()}. When specifying a \code{view_expr} that performs a
slicing or subsetting on a dimension, the equivalent operation will be
applied to the corresponding indexing variable.
}
}
}
\examples{
\dontshow{if (gdal_version_num() >= gdal_compute_version(3, 2, 0)) withAutoprint(\{ # examplesIf}
f_src <- system.file("extdata/byte.nc", package="gdalraster")

## slice along the Y axis with array view
f_dst <- tempfile(fileext = ".nc")
mdim_translate(f, f2, array_specs = "name=Band1,view=[0:10,...]")
(ds <- mdim_as_classic(f_dst, "Band1", 1, 0))

plot_raster(ds, interpolate = FALSE, legend = TRUE,
            main = "Band1[0:10,...]")

dsclose()
\dontshow{deleteDataset(f_dst)}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=mdim_as_classic]{mdim_as_classic()}}, \code{\link[=mdim_info]{mdim_info()}}
}
