% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{plot.OGRFeatureSet}
\alias{plot.OGRFeatureSet}
\title{Plot the geometry column of an \code{OGRFeatureSet}}
\usage{
\method{plot}{OGRFeatureSet}(x, xlab = "x", ylab = "y", main = "", ...)
}
\arguments{
\item{x}{An \code{OGRFeatureSet}.}

\item{xlab}{Title for the x axis.}

\item{ylab}{Title for the y axis.}

\item{main}{The main title (on top).}

\item{...}{Optional arguments passed to \code{wk::wk_plot()}.}
}
\value{
The input, invisibly.
}
\description{
Plot the geometry column of an \code{OGRFeatureSet}
}
