% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_search_paths}
\alias{proj_search_paths}
\title{Get or set search path(s) for PROJ resource files}
\usage{
proj_search_paths(paths = NULL)
}
\arguments{
\item{paths}{Optional character vector containing one or more directory
paths to set.}
}
\value{
A character vector containing the currently used search path(s) for
PROJ resource files. An empty string (\code{""}) is returned if no search paths
are returned by the function \code{OSRGetPROJSearchPaths()} in the GDAL Spatial
Reference System C API.
}
\description{
\code{proj_search_paths()} returns the search path(s) for PROJ resource files,
optionally setting them first.
}
\examples{
proj_search_paths()
}
\seealso{
\code{\link[=proj_version]{proj_version()}}, \code{\link[=proj_networking]{proj_networking()}}
}
