% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{srs_is_same}
\alias{srs_is_same}
\title{Do these two spatial references describe the same system?}
\usage{
srs_is_same(srs1, srs2)
}
\arguments{
\item{srs1}{Character OGC WKT string for a spatial reference system}

\item{srs2}{Character OGC WKT string for a spatial reference system}
}
\value{
Logical. \code{TRUE} if these two spatial references describe the same
system, otherwise \code{FALSE}.
}
\description{
\code{srs_is_same()} returns \code{TRUE} if these two spatial references describe
the same system. This is a wrapper for \code{OSRIsSame()} in the GDAL Spatial
Reference System C API.
}
\examples{
elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
ds <- new(GDALRaster, elev_file, TRUE)
srs_is_same(ds$getProjectionRef(), epsg_to_wkt(26912))
srs_is_same(ds$getProjectionRef(), epsg_to_wkt(5070))
ds$close()
}
